requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.txt -o requirements.txt
absl-py==2.1.0
    # via
    #   -r requirements.txt
    #   dm-control
    #   dm-env
    #   humenv
    #   labmaze
    #   mujoco
annotated-types==0.7.0
    # via
    #   -r requirements.txt
    #   pydantic
anthropic==0.43.0
    # via -r requirements.txt
anyio==4.8.0
    # via
    #   -r requirements.txt
    #   anthropic
    #   httpx
asttokens==3.0.0
    # via
    #   -r requirements.txt
    #   stack-data
blinker==1.9.0
    # via
    #   -r requirements.txt
    #   flask
certifi==2024.12.14
    # via
    #   -r requirements.txt
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.1
    # via
    #   -r requirements.txt
    #   requests
click==8.1.8
    # via
    #   -r requirements.txt
    #   flask
cloudpickle==3.1.1
    # via
    #   -r requirements.txt
    #   gymnasium
contourpy==1.3.1
    # via
    #   -r requirements.txt
    #   matplotlib
cycler==0.12.1
    # via
    #   -r requirements.txt
    #   matplotlib
decorator==5.1.1
    # via
    #   -r requirements.txt
    #   ipython
distro==1.9.0
    # via
    #   -r requirements.txt
    #   anthropic
dm-control==1.0.27
    # via
    #   -r requirements.txt
    #   humenv
dm-env==1.6
    # via
    #   -r requirements.txt
    #   dm-control
dm-tree==0.1.8
    # via
    #   -r requirements.txt
    #   dm-control
    #   dm-env
etils==1.11.0
    # via
    #   -r requirements.txt
    #   mujoco
executing==2.1.0
    # via
    #   -r requirements.txt
    #   stack-data
farama-notifications==0.0.4
    # via
    #   -r requirements.txt
    #   gymnasium
filelock==3.16.1
    # via
    #   -r requirements.txt
    #   huggingface-hub
    #   torch
flask==3.1.0
    # via
    #   -r requirements.txt
    #   flask-cors
flask-cors==5.0.0
    # via -r requirements.txt
fonttools==4.55.3
    # via
    #   -r requirements.txt
    #   matplotlib
fsspec==2024.12.0
    # via
    #   -r requirements.txt
    #   etils
    #   huggingface-hub
    #   torch
glfw==2.8.0
    # via
    #   -r requirements.txt
    #   dm-control
    #   mujoco
gymnasium==1.0.0
    # via
    #   -r requirements.txt
    #   humenv
h11==0.14.0
    # via
    #   -r requirements.txt
    #   httpcore
h5py==3.12.1
    # via
    #   -r requirements.txt
    #   humenv
httpcore==1.0.7
    # via
    #   -r requirements.txt
    #   httpx
httpx==0.28.1
    # via
    #   -r requirements.txt
    #   anthropic
huggingface-hub==0.27.1
    # via -r requirements.txt
humenv @ git+https://github.com/facebookresearch/humenv.git@4f40e9dcea30409d059c82befcdcbee57f00d4a0
    # via -r requirements.txt
idna==3.10
    # via
    #   -r requirements.txt
    #   anyio
    #   httpx
    #   requests
importlib-resources==6.5.2
    # via
    #   -r requirements.txt
    #   etils
inquirerpy==0.3.4
    # via -r requirements.txt
ipython==8.31.0
    # via
    #   -r requirements.txt
    #   mediapy
itsdangerous==2.2.0
    # via
    #   -r requirements.txt
    #   flask
jedi==0.19.2
    # via
    #   -r requirements.txt
    #   ipython
jinja2==3.1.5
    # via
    #   -r requirements.txt
    #   flask
    #   torch
jiter==0.8.2
    # via
    #   -r requirements.txt
    #   anthropic
kiwisolver==1.4.8
    # via
    #   -r requirements.txt
    #   matplotlib
labmaze==1.0.6
    # via
    #   -r requirements.txt
    #   dm-control
lxml==5.3.0
    # via
    #   -r requirements.txt
    #   dm-control
markupsafe==3.0.2
    # via
    #   -r requirements.txt
    #   jinja2
    #   werkzeug
matplotlib==3.10.0
    # via
    #   -r requirements.txt
    #   mediapy
matplotlib-inline==0.1.7
    # via
    #   -r requirements.txt
    #   ipython
mediapy==1.2.2
    # via -r requirements.txt
metamotivo @ git+https://github.com/facebookresearch/metamotivo.git@3b9574709b661d6a60df821bdc30b57584bdfc64
    # via -r requirements.txt
mpmath==1.3.0
    # via
    #   -r requirements.txt
    #   sympy
mujoco==3.2.7
    # via
    #   -r requirements.txt
    #   dm-control
    #   humenv
networkx==3.4.2
    # via
    #   -r requirements.txt
    #   torch
numpy==2.2.1
    # via
    #   -r requirements.txt
    #   contourpy
    #   dm-control
    #   dm-env
    #   gymnasium
    #   h5py
    #   humenv
    #   labmaze
    #   matplotlib
    #   mediapy
    #   mujoco
    #   opencv-python
    #   pot
    #   scipy
opencv-python==4.10.0.84
    # via -r requirements.txt
packaging==24.2
    # via
    #   -r requirements.txt
    #   huggingface-hub
    #   humenv
    #   matplotlib
parso==0.8.4
    # via
    #   -r requirements.txt
    #   jedi
pexpect==4.9.0
    # via
    #   -r requirements.txt
    #   ipython
pfzy==0.3.4
    # via
    #   -r requirements.txt
    #   inquirerpy
pillow==11.1.0
    # via
    #   -r requirements.txt
    #   matplotlib
    #   mediapy
pot==0.9.5
    # via -r requirements.txt
prompt-toolkit==3.0.48
    # via
    #   -r requirements.txt
    #   inquirerpy
    #   ipython
protobuf==5.29.3
    # via
    #   -r requirements.txt
    #   dm-control
ptyprocess==0.7.0
    # via
    #   -r requirements.txt
    #   pexpect
pure-eval==0.2.3
    # via
    #   -r requirements.txt
    #   stack-data
pydantic==2.10.5
    # via
    #   -r requirements.txt
    #   anthropic
pydantic-core==2.27.2
    # via
    #   -r requirements.txt
    #   pydantic
pygments==2.19.1
    # via
    #   -r requirements.txt
    #   ipython
pyopengl==3.1.7
    # via
    #   -r requirements.txt
    #   dm-control
    #   mujoco
pyparsing==3.2.1
    # via
    #   -r requirements.txt
    #   dm-control
    #   matplotlib
python-dateutil==2.9.0.post0
    # via
    #   -r requirements.txt
    #   matplotlib
pyyaml==6.0.2
    # via
    #   -r requirements.txt
    #   huggingface-hub
requests==2.32.3
    # via
    #   -r requirements.txt
    #   dm-control
    #   huggingface-hub
safetensors==0.5.2
    # via
    #   -r requirements.txt
    #   metamotivo
scipy==1.15.1
    # via
    #   -r requirements.txt
    #   dm-control
    #   pot
setuptools==75.8.0
    # via
    #   -r requirements.txt
    #   dm-control
    #   labmaze
six==1.17.0
    # via
    #   -r requirements.txt
    #   python-dateutil
sniffio==1.3.1
    # via
    #   -r requirements.txt
    #   anthropic
    #   anyio
stack-data==0.6.3
    # via
    #   -r requirements.txt
    #   ipython
sympy==1.13.1
    # via
    #   -r requirements.txt
    #   torch
torch==2.5.1
    # via
    #   -r requirements.txt
    #   metamotivo
tqdm==4.67.1
    # via
    #   -r requirements.txt
    #   dm-control
    #   huggingface-hub
traitlets==5.14.3
    # via
    #   -r requirements.txt
    #   ipython
    #   matplotlib-inline
typing-extensions==4.12.2
    # via
    #   -r requirements.txt
    #   anthropic
    #   anyio
    #   etils
    #   gymnasium
    #   huggingface-hub
    #   ipython
    #   pydantic
    #   pydantic-core
    #   torch
urllib3==2.3.0
    # via
    #   -r requirements.txt
    #   requests
wcwidth==0.2.13
    # via
    #   -r requirements.txt
    #   prompt-toolkit
websockets==14.1
    # via -r requirements.txt
werkzeug==3.1.3
    # via
    #   -r requirements.txt
    #   flask
zipp==3.21.0
    # via
    #   -r requirements.txt
    #   etils
